float specs

+++ src/main/java/cz/vse/potravinyBEZ/domain/product/CreateProductRequest.java
@@ -19,16 +19,16 @@ public class CreateProductRequest {
     private String descriptionShort;
     @Length(min = 2, max = 1024)
     private String descriptionLong;
-    private int weight;
+    private float weight;
     @Length(min = 2, max = 1024)
     private String ingredients;
     @Length(min = 2, max = 45)
     private String producer;
-    private int energy;
-    private int fat;
-    private int carbohydrate;
-    private int sugars;
-    private int fibre;
-    private int protein;
-    private int salt;
+    private float energy;
+    private float fat;
+    private float carbohydrate;
+    private float sugars;
+    private float fibre;
+    private float protein;
+    private float salt;
 }
